+++ src/Config/Option/UrlShortener/GeoLiteLicenseKeyConfigOption.php
@@ -15,13 +15,11 @@ class GeoLiteLicenseKeyConfigOption extends BaseConfigOption
         return ['geolite2', 'license_key'];
     }
 
-    public function ask(StyleInterface $io, PathCollection $currentOptions): string
+    public function ask(StyleInterface $io, PathCollection $currentOptions): ?string
     {
-        // TODO For Shlink 3.0, this option should be mandatory. The default value should be removed
         return $io->ask(
-            'Provide a GeoLite2 license key. (Leave empty to use default one, but it is '
-            . '<fg=green;options=bold>strongly recommended</> to get your own. '
-            . 'Go to https://shlink.io/documentation/geolite-license-key to know how to get it)',
-        ) ?? 'G4Lm0C60yJsnkdPi'; // Deprecated
+            'Provide a GeoLite2 license key. Leave empty to disable geolocation. '
+            . '(Go to https://shlink.io/documentation/geolite-license-key to know how to generate it)',
+        );
     }
 }